custom-elements/instantiating-custom-elements/extensions-to-document-interface/created-callback-test-004.html

<!DOCTYPE html>
<html>
<head>
<title>Created callback should be called if createElementNS is called</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Created callback should be called if createElementNS is called">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

function doTest(element) {
    var doc = newHTMLDocument();
    
    var obj = doc.createElement(element);
    
    var name = 'x-foo-' + element;
    
    var proto = Object.create(obj.constructor.prototype);
    proto.createdCallback = function () {
        callbackCalled = true;
    }    
    
    var XFoo = doc.register(name, {prototype: proto, extends: element});
    
    var xfoo = doc.createElementNS(HTML_NAMESPACE, element, name);
        
    assert_true(callbackCalled, 'Callback created was not called');
    
}


var parameters = [];

for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
    if (HTML5_DOCUMENT_ELEMENTS.indexOf(HTML5_ELEMENTS[i]) == -1) {
        parameters[i] = ['Custom element type ' + HTML5_ELEMENTS[i], HTML5_ELEMENTS[i]];
    }
}

generate_tests(doTest, parameters, 
        'If typeExtension presents in createElement call then type is a type extension');

</script>
</body>
</html>
